+++ ucc.frontend/src/main/java/org/universAAL/ucc/service/manager/Activator.java
@@ -237,10 +237,12 @@ public class Activator implements BundleActivator {
 	private void deleteFiles(File path) {
 		File[] files = path.listFiles();
 		for (File del : files) {
-			if (del.isDirectory()) {
+			if (del.isDirectory()
+					&& !del.getPath().substring(del.getPath().lastIndexOf(".") + 1)
+							.equals("usrv")) {
 				deleteFiles(del);
 			}
-			if (!del.getPath().substring(del.getPath().indexOf(".") + 1)
+			if (!del.getPath().substring(del.getPath().lastIndexOf(".") + 1)
 					.equals("usrv"))
 				del.delete();
 		}
